requirements.txt
opencv-python
numpy
scikit-learn
Pillow
picamera
